refactor: move aviation subsystem (client, tracker, commands) into aviation/ module
