+++ src/aviation/mod.rs
@@ -1,3 +1,8 @@
+pub mod commands;
+pub mod tracker;
+
+pub use tracker::{FlightIdentifier, TrackerCommand, run_flight_tracker};
+
 use eyre::{Result, WrapErr};
 use serde::Deserialize;
 use std::collections::HashMap;
@@ -27,7 +32,7 @@ const UP_CONE_REFERENCE_ALT_FT: f64 = 35_000.0;
 
 // --- PLZ Lookup ---
 
-const PLZ_DATA: &str = include_str!("../data/plz.csv");
+const PLZ_DATA: &str = include_str!("../../data/plz.csv");
 
 fn plz_table() -> &'static HashMap<&'static str, (f64, f64)> {
     static TABLE: OnceLock<HashMap<&'static str, (f64, f64)>> = OnceLock::new();
@@ -62,7 +67,7 @@ fn is_valid_plz(plz: &str) -> bool {
 
 // --- Airport Lookup ---
 
-const AIRPORT_DATA: &str = include_str!("../data/airports.csv");
+const AIRPORT_DATA: &str = include_str!("../../data/airports.csv");
 
 struct AirportData {
     by_icao: HashMap<String, (f64, f64, String)>,
@@ -121,7 +126,7 @@ pub(crate) fn iata_to_coords(code: &str) -> Option<(f64, f64, &'static str)> {
 
 // --- Airline IATA-to-ICAO Lookup ---
 
-const AIRLINE_DATA: &str = include_str!("../data/airlines.csv");
+const AIRLINE_DATA: &str = include_str!("../../data/airlines.csv");
 
 /// Returns the static IATA→ICAO airline code table (lazy-initialized).
 fn airline_table() -> &'static HashMap<&'static str, &'static str> {
